Update code_verification.py
- We have added `self.alias_map = {}` in the `__init__` method:
This adds a new dictionary to the instance of `FunctionCallValidator` called `alias_map`. This dictionary will be used to track aliases that correspond to blocked function names.

- Adding the `visit_Assign` method:

def visit_Assign(self, node):


This method is defined to handle assignment nodes in the abstract syntax tree (AST). Whenever an assignment is encountered in the source code, this method is called with the corresponding node.


for target in node.targets:

We iterate over all the targets of the assignment because in Python, it is possible to assign a value to multiple variables at the same time (for example, `a = b = 5`).


if isinstance(target, ast.Name):

This checks if the target of the assignment is a variable (and not a list or another data structure).

if isinstance(node.value, ast.Name):

We then check if the assigned value is also a variable. If so, this could indicate an alias, for example `a = some_function`.


if node.value.id in self.blocked_functions:

We look to see if the assigned value (for example, the function name) is in the list of blocked functions.

self.alias_map[target.id] = node.value.id:

If the above condition is true, this means the user has created an alias for a blocked function, and this alias is recorded in `alias_map`.

- Modifying the `visit_Call` method:

if isinstance(node.func, ast.Name):

This is a check to see if the called function node is a function by its direct name and not an attribute or a method.


if function_name in self.blocked_functions or function_name in self.alias_map:

This condition now checks two things: whether the function name is blocked or if it exists in `alias_map` as an alias for a blocked function. This is where TaskWeaver detects the use of aliases for blocked functions.


self.errors.append(...):

If either of the above conditions is true, an error message is added to the list of errors.


return False:

If a blocked function or its alias is detected, the method returns False, indicating that the validation has failed.


return True:

If no blocked function or alias is detected, the method returns True, indicating that this part of the code is valid.